+++ Source/SwiftLintFramework/LintOrAnalyzeCommand.swift
@@ -38,6 +38,7 @@ package struct LintOrAnalyzeOptions {
     let forceExclude: Bool
     let useExcludingByPrefix: Bool
     let useScriptInputFiles: Bool
+    let useScriptInputFileLists: Bool
     let benchmark: Bool
     let reporter: String?
     let baseline: String?
@@ -65,6 +66,7 @@ package struct LintOrAnalyzeOptions {
                  forceExclude: Bool,
                  useExcludingByPrefix: Bool,
                  useScriptInputFiles: Bool,
+                 useScriptInputFileLists: Bool,
                  benchmark: Bool,
                  reporter: String?,
                  baseline: String?,
@@ -91,6 +93,7 @@ package struct LintOrAnalyzeOptions {
         self.forceExclude = forceExclude
         self.useExcludingByPrefix = useExcludingByPrefix
         self.useScriptInputFiles = useScriptInputFiles
+        self.useScriptInputFileLists = useScriptInputFileLists
         self.benchmark = benchmark
         self.reporter = reporter
         self.baseline = baseline